+++ oshi-core/src/main/java/oshi/driver/windows/perfmon/MemoryInformation.java
@@ -73,8 +73,6 @@ public final class MemoryInformation {
      * @return Page swap counters for memory.
      */
     public static Map<PageSwapProperty, Long> queryPageSwaps() {
-        PerfCounterQuery<PageSwapProperty> memoryPerfCounters = new PerfCounterQuery<>(PageSwapProperty.class, MEMORY,
-                WIN32_PERF_RAW_DATA_PERF_OS_MEMORY);
-        return memoryPerfCounters.queryValues();
+        return PerfCounterQuery.queryValues(PageSwapProperty.class, MEMORY, WIN32_PERF_RAW_DATA_PERF_OS_MEMORY);
     }
 }