+++ oshi-core/src/main/java/oshi/driver/windows/perfmon/PhysicalDisk.java
@@ -76,11 +76,7 @@ public final class PhysicalDisk {
      * @return Performance Counters for physical disks.
      */
     public static Pair<List<String>, Map<PhysicalDiskProperty, List<Long>>> queryDiskCounters() {
-        PerfCounterWildcardQuery<PhysicalDiskProperty> physicalDiskPerfCounters = new PerfCounterWildcardQuery<>(
-                PhysicalDiskProperty.class, PHYSICAL_DISK,
+        return PerfCounterWildcardQuery.queryInstancesAndValues(PhysicalDiskProperty.class, PHYSICAL_DISK,
                 WIN32_PERF_RAW_DATA_PERF_DISK_PHYSICAL_DISK_WHERE_NOT_NAME_TOTAL);
-        Map<PhysicalDiskProperty, List<Long>> values = physicalDiskPerfCounters.queryValuesWildcard();
-        List<String> instances = physicalDiskPerfCounters.getInstancesFromLastQuery();
-        return new Pair<>(instances, values);
     }
 }