+++ oshi-core/src/main/java/oshi/software/os/linux/LinuxOperatingSystem.java
@@ -155,7 +155,7 @@ public class LinuxOperatingSystem extends AbstractOperatingSystem {
 
     @Override
     public List<OSSession> getSessions() {
-        return Collections.unmodifiableList(Who.queryUtxent());
+        return Collections.unmodifiableList(USE_WHO_COMMAND ? super.getSessions() : Who.queryUtxent());
     }
 
     @Override