+++ oshi-core/src/main/java/oshi/software/os/mac/MacOperatingSystem.java
@@ -184,7 +184,7 @@ public class MacOperatingSystem extends AbstractOperatingSystem {
 
     @Override
     public List<OSSession> getSessions() {
-        return Collections.unmodifiableList(Who.queryUtxent());
+        return Collections.unmodifiableList(USE_WHO_COMMAND ? super.getSessions() : Who.queryUtxent());
     }
 
     @Override